+++ apps/cold-api/src/platform/modules/resources/organizations/facilities/footprints/footprints.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { FootprintsService } from './footprints.service';
+import { FootprintsController } from './footprints.controller';
+
+@Module({
+  controllers: [FootprintsController],
+  providers: [FootprintsService],
+})
+export class FootprintsModule {}