+++ apps/cold-api/src/platform/modules/resources/scopes/dto/create-scope.dto.ts
@@ -1,0 +1,7 @@
+export class CreateScopeDto {
+  ghg_category: number;
+  ghg_subcategory?: number | null;
+  label: string;
+  name: string;
+  organization_id?: string;
+}