+++ libs/nest/prisma/migrations/20240405203722_rename_scope_and_subscope/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "emission_scopes"
+  RENAME COLUMN "scope" TO "ghg_category";
+ALTER TABLE "emission_scopes"
+  RENAME COLUMN "sub_scope" TO "ghg_subcategory";